BlackBerry: update plugin key
The plugin key for both the Generic backend and the BlackBerry backend
were set to "notused". This ultimately caused the Generic backend not
to be loaded, which is why QTiltSensor wasn't being exposed. The duplicate
keys caused a collision in a QMap in QFactoryLoader, making it look
like the generic backend had been created, when it hadn't.

Tested on a BlackBerry Z10.

Change-Id: Ie5e40586751e7c3b9683a94a39b07efd4fc7fc82
Reviewed-by: Bernd Weimer <bweimer@blackberry.com>
Reviewed-by: Fabian Bumberger <fbumberger@rim.com>